Moved test cases for issue #2191 into OperatorReplayTest
